Piping and hooks
Added a Pipe system to allow the UI to talk with the mod hook

Added in code that will allow the mod hook to get ahold of the main
server loop of the Server

Made the log output come out in HTML to allow for nicer output style
and to allow me to change the colors on the whim when required.

Got muting, kicking and maybe banning working. Need to look into a few
more touches before releasing it

Will need to speak with server hosts to see what they would like access
to and how they would like to access such areas of the Staxel Server.